requirements.txt
# ------------------------------
# Core Inference Dependencies
# ------------------------------
torch>=2.1.0
transformers>=4.40.0
peft>=0.7.0
scikit-image>=0.22.0
numpy>=1.26.0
Pillow>=10.0.0
matplotlib>=3.8.0

# ------------------------------
# Backend (Flask API)
# ------------------------------
flask>=3.0.0
flask-cors>=4.0.0

# ------------------------------
# Utilities
# ------------------------------
tqdm>=4.66.0
